Move block into folder
